Refactored IMSC reader and writer
