+++ src/main/python/ttconv/imsc/style_properties.py
@@ -28,7 +28,7 @@
 import inspect
 import math
 import typing
-import ttconv.imsc.namespaces as ns
+import ttconv.imsc.namespaces as xml_ns
 import ttconv.style_properties as styles
 import ttconv.imsc.utils as utils
 
@@ -53,7 +53,7 @@ class StyleProperties:
   class BackgroundColor(StyleProperty):
     '''Corresponds to tts:backgroundColor.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "backgroundColor"
     model_prop = styles.StyleProperties.BackgroundColor
 
@@ -64,7 +64,7 @@ class StyleProperties:
   class Color(StyleProperty):
     '''Corresponds to tts:color.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "color"
     model_prop = styles.StyleProperties.Color
 
@@ -76,7 +76,7 @@ class StyleProperties:
   class Direction(StyleProperty):
     '''Corresponds to tts:direction.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "direction"
     model_prop = styles.StyleProperties.Direction
 
@@ -89,7 +89,7 @@ class StyleProperties:
     '''Corresponds to tts:disparity.
     '''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "disparity"
     model_prop = styles.StyleProperties.Disparity
 
@@ -101,7 +101,7 @@ class StyleProperties:
   class Display(StyleProperty):
     '''Corresponds to tts:display.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "display"
     model_prop = styles.StyleProperties.Display
 
@@ -113,7 +113,7 @@ class StyleProperties:
   class DisplayAlign(StyleProperty):
     '''Corresponds to tts:displayAlign.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "displayAlign"
     model_prop = styles.StyleProperties.DisplayAlign
 
@@ -125,7 +125,7 @@ class StyleProperties:
     '''Corresponds to tts:extent.
     '''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "extent"
     model_prop = styles.StyleProperties.Extent
 
@@ -152,7 +152,7 @@ class StyleProperties:
   class FillLineGap(StyleProperty):
     '''Corresponds to itts:fillLineGap.'''
 
-    ns = ns.ITTS
+    ns = xml_ns.ITTS
     local_name = "fillLineGap"
     model_prop = styles.StyleProperties.FillLineGap
 
@@ -163,7 +163,7 @@ class StyleProperties:
   class FontFamily(StyleProperty):
     '''Corresponds to tts:fontFamily.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "fontFamily"
     model_prop = styles.StyleProperties.FontFamily
 
@@ -178,7 +178,7 @@ class StyleProperties:
   class FontSize(StyleProperty):
     '''Corresponds to tts:fontSize.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "fontSize"
     model_prop = styles.StyleProperties.FontSize
 
@@ -190,7 +190,7 @@ class StyleProperties:
   class FontStyle(StyleProperty):
     '''Corresponds to tts:fontStyle.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "fontStyle"
     model_prop = styles.StyleProperties.FontStyle
 
@@ -202,7 +202,7 @@ class StyleProperties:
   class FontWeight(StyleProperty):
     '''Corresponds to tts:fontWeight.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "fontWeight"
     model_prop = styles.StyleProperties.FontWeight
 
@@ -214,7 +214,7 @@ class StyleProperties:
   class LineHeight(StyleProperty):
     '''tts:lineHeight'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "lineHeight"
     model_prop = styles.StyleProperties.LineHeight
 
@@ -227,7 +227,7 @@ class StyleProperties:
   class LinePadding(StyleProperty):
     '''Corresponds to ebutts:linePadding.'''
 
-    ns = ns.EBUTTS
+    ns = xml_ns.EBUTTS
     local_name = "linePadding"
     model_prop = styles.StyleProperties.LinePadding
 
@@ -242,7 +242,7 @@ class StyleProperties:
   class LuminanceGain(StyleProperty):
     '''Corresponds to tts:luminanceGain.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "luminanceGain"
     model_prop = styles.StyleProperties.LuminanceGain
 
@@ -254,7 +254,7 @@ class StyleProperties:
   class MultiRowAlign(StyleProperty):
     '''Corresponds to ebutts:multiRowAlign.'''
 
-    ns = ns.EBUTTS
+    ns = xml_ns.EBUTTS
     local_name = "multiRowAlign"
     model_prop = styles.StyleProperties.MultiRowAlign
 
@@ -266,7 +266,7 @@ class StyleProperties:
   class Opacity(StyleProperty):
     '''Corresponds to tts:opacity.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "opacity"
     model_prop = styles.StyleProperties.Opacity
 
@@ -278,7 +278,7 @@ class StyleProperties:
   class Origin(StyleProperty):
     '''Corresponds to tts:origin.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "origin"
     model_prop = styles.StyleProperties.Origin
 
@@ -308,7 +308,7 @@ class StyleProperties:
   class Overflow(StyleProperty):
     '''Corresponds to tts:overflow.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "overflow"
     model_prop = styles.StyleProperties.Overflow
 
@@ -320,7 +320,7 @@ class StyleProperties:
   class Padding(StyleProperty):
     '''Corresponds to tts:padding.'''
     
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "padding"
     model_prop = styles.StyleProperties.Padding
 
@@ -359,7 +359,7 @@ class StyleProperties:
   class Position(StyleProperty):
     '''Corresponds to tts:position.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "position"
     model_prop = styles.StyleProperties.Origin
 
@@ -374,7 +374,7 @@ class StyleProperties:
   class RubyAlign(StyleProperty):
     '''Corresponds to tts:rubyAlign.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "rubyAlign"
     model_prop = styles.StyleProperties.RubyAlign
 
@@ -386,7 +386,7 @@ class StyleProperties:
   class RubyPosition(StyleProperty):
     '''Corresponds to tts:rubyPosition.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "rubyPosition"
     model_prop = styles.StyleProperties.RubyPosition
   
@@ -397,7 +397,7 @@ class StyleProperties:
   class RubyReserve(StyleProperty):
     '''Corresponds to tts:rubyReserve.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "rubyReserve"
     model_prop = styles.StyleProperties.RubyReserve
 
@@ -422,7 +422,7 @@ class StyleProperties:
   class Shear(StyleProperty):
     '''Corresponds to tts:shear.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "shear"
     model_prop = styles.StyleProperties.Shear
 
@@ -440,7 +440,7 @@ class StyleProperties:
   class TextAlign(StyleProperty):
     '''Corresponds to tts:textAlign.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "textAlign"
     model_prop = styles.StyleProperties.TextAlign
   
@@ -464,7 +464,7 @@ class StyleProperties:
   class TextCombine(StyleProperty):
     '''Corresponds to tts:textCombine.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "textCombine"
     model_prop = styles.StyleProperties.TextCombine
 
@@ -476,7 +476,7 @@ class StyleProperties:
   class TextDecoration(StyleProperty):
     '''Corresponds to tts:textDecoration.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "textDecoration"
     model_prop = styles.StyleProperties.TextDecoration
 
@@ -515,7 +515,7 @@ class StyleProperties:
   class TextEmphasis(StyleProperty):
     '''Corresponds to tts:textEmphasis.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "textEmphasis"
     model_prop = styles.StyleProperties.TextEmphasis
 
@@ -570,7 +570,7 @@ class StyleProperties:
   class TextOutline(StyleProperty):
     '''Corresponds to tts:textOutline.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "textOutline"
     model_prop = styles.StyleProperties.TextOutline
 
@@ -597,7 +597,7 @@ class StyleProperties:
   class TextShadow(StyleProperty):
     '''Corresponds to tts:textShadow.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "textShadow"
     model_prop = styles.StyleProperties.TextShadow
 
@@ -652,7 +652,7 @@ class StyleProperties:
   class UnicodeBidi(StyleProperty):
     '''Corresponds to tts:unicodeBidi.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "unicodeBidi"
     model_prop = styles.StyleProperties.UnicodeBidi
 
@@ -663,7 +663,7 @@ class StyleProperties:
   class Visibility(StyleProperty):
     '''Corresponds to tts:visibility.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "visibility"
     model_prop = styles.StyleProperties.Visibility
     
@@ -674,7 +674,7 @@ class StyleProperties:
   class WrapOption(StyleProperty):
     '''Corresponds to tts:wrapOption.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "wrapOption"
     model_prop = styles.StyleProperties.WrapOption
 
@@ -685,7 +685,7 @@ class StyleProperties:
   class WritingMode(StyleProperty):
     '''Corresponds to tts:writingMode.'''
 
-    ns = ns.TTS
+    ns = xml_ns.TTS
     local_name = "writingMode"
     model_prop = styles.StyleProperties.WritingMode
