+++ src/main/python/ttconv/imsc/writer.py
